src/test/java/mancala/BoardTest.java
/***************
 * This test serves the sole purpose of verifying the existence of methods in the class.
 * It does not validate the types of expected return values, but it does confirm the presence of a return type if expected.
 * Please note that this form of testing is not suitable for comprehensive unit testing.
 * It has been designed solely as a tool to identify missing methods within your implementation.
 */

import mancala.*;
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;

public class BoardTest {

    @Test
    public void testCaptureStones() {
        Board board = new Board(); // Instantiate Board
        int stonesCaptured = board.captureStones(4);

        assertEquals(0, stonesCaptured);
    }

    @Test
    public void testDistributeStones() {
        Board board = new Board(); // Instantiate Board
        int lastPit = board.distributeStones(5);
        assertTrue(lastPit >= 0 && lastPit < 12);
    }
    
    @Test
    public void testGetNumStones() {
        Board board = new Board(); // Instantiate Board
        int stoneCount = board.getNumStones(5);

        assertEquals(4, stoneCount);
    }

    @Test
    public void testIsSideEmpty() {
        Board board = new Board(); // Instantiate Board
        boolean isSideEmpty = board.isSideEmpty(8);

        assertFalse(isSideEmpty);
    }

    @Test
    public void testResetBoard() {
        Board board = new Board(); // Instantiate Board
        board.resetBoard();

        // Add additional assertions based on reset logic
    }

    @Test
    public void testRegisterPlayers() {
        Board board = new Board(); // Instantiate Board
        Player playerOne = new Player("Alice");
        Player playerTwo = new Player("Bob");
        board.registerPlayers(playerOne, playerTwo);

        // Add additional assertions based on registration logic
    }
}
